separate json_ai/codegen modules
